pyproject.toml
[build-system]
requires = ["setuptools>=59.5.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pointnext"
version = "0.1.0"
description = "PointNeXt: Revisiting PointNet++ with Improved Training and Scaling Strategies"
requires-python = ">=3.10,<3.11"
dependencies = [
    # Local C++ extensions (resolved via workspace)
    "pointnet2-cuda",
    "pointops",
    "chamfer",
    "emd-ext",
    "grid-subsampling",
    
    # Core dependencies
    "torchvision>=0.19.0",
    "torch-scatter",
    "numpy<2.0",
    
    # ML utilities
    "scikit-learn==1.0.2",
    "deepspeed",
    "wandb",
    "tensorboard==2.8.0",
    
    # Data handling
    "pandas",
    "gdown",
    "PyYAML==6.0",
    "easydict==1.9",
    "h5py>=3.6.0",
    
    # Visualization
    "matplotlib==3.5.1",
    "pyvista",
    
    # Training utilities
    "tqdm==4.62.3",
    "pillow",
    "protobuf==3.19.4",
    "multimethod==1.7",
    "ninja==1.10.2.3",
    "Cython==0.29.28",
    "setuptools==59.5.0",
    
    # Optional
    "pickleshare==0.7.5",
    "termcolor==1.1.0",
    "shortuuid",
    
    # Docs
    "mkdocs-material",
    "mkdocs-awesome-pages-plugin",
    "mdx_truly_sane_lists",
]

[tool.setuptools.packages.find]
include = ["openpoints*", "examples*"]
